Update landing to include query templates
